+++ tests/testthat/test-writeData.R
@@ -35,3 +35,38 @@ test_that("writeData() forces evaluation of x (#264)", {
   options(op)
   file.remove(wbfile)
 })
+
+
+test_that("as.character.formula() works [312]", {
+  form <- y ~ x1 * x2 + x3
+  expect_identical(
+    as.character.default(form),
+    openxlsx::as.character.formula(form)
+  )
+  
+  skip_if_not_installed(
+    "formula.tools",
+    "tests specifically for as.character.formula conflict"
+  )
+  
+  foo <- function() {
+    wb <- openxlsx::buildWorkbook(
+      data.frame(
+        x = structure("A2 + B2", class = c("character", "formula")),
+        stringsAsFactors = FALSE
+      )
+    )
+    as.list(wb$worksheets[[1]]$sheet_data)
+  }
+  
+  before <- foo()
+  # don't required the "require" function for deps  check
+  match.fun("require")("formula.tools", character.only = TRUE)
+  middle <- foo()
+  detach("package:formula.tools", character.only = TRUE, force = TRUE)
+  end <- foo()
+  
+  expect_identical(before, middle, ignore.environment = TRUE)
+  expect_identical(before, end,    ignore.environment = TRUE)
+})
+